Fix docker issues
For some reason the import of svgs does not work when we also import react-arborist, otherwise it does. Another issue was that the ssh config was not in the database folder, but it is still not properly tested yet. And we also provided initial version of the config for backend (we will have to change the git one, because there are more providers).