Add force reload for diff tree data and force it after editor save
This solves the issue that we need to update the diff tree data both in the editor and on the backend. So we just force store it to backend and reload on frontend. It would be better to update the backend merge state partially using some create, remove, modify commands, but that is not as easy as it sounds. Also we have to update the backend anyways when creating new resources, because we need the backend to generate iri.